requirements.txt
Flask
Flask-Login
SQLAlchemy
requests
python-telegram-bot
python-dotenv
matplotlib
Werkzeug